Session4/Session4-Base/requirements.txt
flask
flask-socketio
torch
torchvision
pillow
numpy
python-multipart
tqdm>=4.65.0